requirements.txt
apache-airflow==2.10.0
requests
pandas
psycopg2-binary
python-dotenv